requirements.txt
faiss-cpu
sentence-transformers
transformers
torch
vllm
fastapi
uvicorn
pyyaml
numpy
accelerate
optimum
auto-gptq
huggingface-hub
